fix: improve editor handling
